refactor: split user_types.py into domain-specific package
Split the 549-line user_types.py into three modules: data.py (core
types and data structures), certificates.py (certificate function
types), and callables.py (callable protocols and configs). The
__init__.py re-exports everything for backward compatibility.
